update zk verify template
